+++ MAPI/StubUtils.cpp
@@ -324,7 +324,7 @@ namespace mapistub
 		{
 			auto hRes = S_OK;
 			buf.resize(buf.size() + MAX_PATH);
-			EC_D(copied, ::GetSystemDirectoryW(&buf.at(0), buf.size()));
+			EC_D(copied, ::GetSystemDirectoryW(&buf.at(0), static_cast<DWORD>(buf.size())));
 		} while (copied >= buf.size());
 
 		buf.resize(copied);
